# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LocalMind API",
        "description": "This is an API documentation for LocalMind.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3535/api",
            "description": "LocalMind API Server"
        }
    ],
    "paths": {
...
        "/questions": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "Get all questions",
                "operationId": "7a2338bdf2544317e8bc9296c65f1f4d",
                "responses": {
                    "200": {
                        "description": "Questions retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Indicates if the questions were retrieved successfully",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Question"
                                        },
                                        "message": {
                                            "description": "A success message",
                                            "type": "string",
                                            "example": "Questions retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Indicates if the questions were retrieved successfully",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "An error message",
                                            "type": "string",
                                            "example": "Something went wrong, please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Questions"
                ],
                "summary": "Create a new question",
                "operationId": "a12f796f2fdc772ce53ecadbad7e73f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "The title of the question",
                                        "type": "string",
                                        "example": "This is my question title."
                                    },
                                    "description": {
                                        "description": "The description of the question",
                                        "type": "string",
                                        "example": "This is my question description."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Question created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Indicates if the question was created successfully",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Question"
                                        },
                                        "message": {
                                            "description": "A success message",
                                            "type": "string",
                                            "example": "Question created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Indicates if the question was created successfully",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "An error message",
                                            "type": "string",
                                            "example": "Something went wrong, please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/questions/{question}": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "Get a specific question",
                "operationId": "7b2a97ff28923364deb6494790cfe1b8",
                "parameters": [
                    {
                        "name": "question",
                        "in": "path",
                        "description": "The ID of the question to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Indicates if the question was retrieved successfully",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Question"
                                        },
                                        "message": {
                                            "description": "A success message",
                                            "type": "string",
                                            "example": "Question retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Indicates if the question was retrieved successfully",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "An error message",
                                            "type": "string",
                                            "example": "Something went wrong, please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Questions"
                ],
                "summary": "Update a specific question",
                "operationId": "0b3af88f3ffb460e4a04a3d923c04b6e",
                "parameters": [
                    {
                        "name": "question",
                        "in": "path",
                        "description": "The ID of the question to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "The updated title of the question.",
                                        "type": "string",
                                        "example": "Updated question title."
                                    },
                                    "description": {
                                        "description": "The updated description of the question.",
                                        "type": "string",
                                        "example": "Updated question description."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Question updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Indicates if the question was updated successfully",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Question"
                                        },
                                        "message": {
                                            "description": "A success message",
                                            "type": "string",
                                            "example": "Question updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Indicates if the question was updated successfully",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "An error message",
                                            "type": "string",
                                            "example": "Something went wrong, please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Questions"
                ],
                "summary": "Delete a specific question",
                "operationId": "a2579f150ed0a9aa690cc42b0b5fec5c",
                "parameters": [
                    {
                        "name": "question",
                        "in": "path",
                        "description": "The ID of the question to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question deleted successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Indicates if the question was deleted successfully",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "A success message",
                                            "type": "string",
                                            "example": "Question deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Indicates if the question was deleted successfully",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "An error message",
                                            "type": "string",
                                            "example": "Something went wrong, please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Question": {
                "title": "Question",
                "description": "Question model",
                "required": [
                    "title",
                    "description",
                    "user_id"
                ],
                "properties": {
                    "id": {
                        "description": "The ID of the question",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "description": "The title of the question",
                        "type": "string",
                        "example": "What is Laravel?"
                    },
                    "description": {
                        "description": "The description of the question",
                        "type": "string",
                        "example": "Laravel is a web application framework with expressive, elegant syntax."
                    },
                    "user_id": {
                        "description": "The ID of the user who asked the question",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}